ReviewSense-AI/requirements.txt
# Core Deep Learning (Mac optimized)
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Hugging Face Ecosystem
transformers==4.36.2
datasets==2.16.1
accelerate==0.25.0
peft==0.7.1
trl==0.7.10
sentencepiece==0.1.99
protobuf==4.25.1

# Evaluation Metrics
evaluate==0.4.1
rouge-score==0.1.2
sacrebleu==2.3.1
nltk==3.8.1

# Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Progress Bars
tqdm==4.66.1

# Logging & Experiment Tracking
wandb==0.16.1
tensorboard==2.15.1

# Utilities
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.28.0

# Deployment
streamlit==1.29.0
gradio==4.13.0

# API clients
huggingface-hub==0.20.1

# JSON handling
ujson==5.9.0
